+++ src/pal-api.c
@@ -27,8 +27,8 @@
 #include <string.h>
 
 enum PalReturnCode pal_api_init(struct PalHandler *hpal,
-                                size_t tx_capacity,
-                                size_t max_msg_size,
+                                uint32_t tx_capacity,
+                                uint32_t max_msg_size,
                                 pal_send_fn send,
                                 void (*cs_enter)(void),
                                 void (*cs_exit)(void),
@@ -40,17 +40,17 @@ enum PalReturnCode pal_api_init(struct PalHandler *hpal,
         return PAL_RC_INVALID_PARAM;
 
     hpal->send = send;
-    ring_buffer_api_init(&hpal->tx_queue, sizeof(size_t) + max_msg_size , tx_capacity, cs_enter, cs_exit, arena);
-    hpal->add_to_tx_msg = (uint8_t *)arena_allocator_api_alloc(arena, sizeof(size_t) + max_msg_size );
-    hpal->exec_tx_msg = (uint8_t *)arena_allocator_api_alloc(arena, sizeof(size_t) + max_msg_size );
+    ring_buffer_api_init(&hpal->tx_queue, sizeof(uint32_t) + max_msg_size , tx_capacity, cs_enter, cs_exit, arena);
+    hpal->add_to_tx_msg = (uint8_t *)arena_allocator_api_alloc(arena, sizeof(uint32_t) + max_msg_size );
+    hpal->exec_tx_msg = (uint8_t *)arena_allocator_api_alloc(arena, sizeof(uint32_t) + max_msg_size );
     hpal->max_msg_size = max_msg_size;
     return PAL_RC_OK;
 }
 
-enum PalReturnCode pal_api_add_to_tx_queue(struct PalHandler *hpal, void *buff,size_t size) {
+enum PalReturnCode pal_api_add_to_tx_queue(struct PalHandler *hpal, void *buff,uint32_t size) {
     if (hpal == NULL || buff == NULL)
         return PAL_RC_NULL_PTR;
-    
+
     if (size <= 0)
         return PAL_RC_INVALID_PARAM;
 
@@ -60,8 +60,8 @@ enum PalReturnCode pal_api_add_to_tx_queue(struct PalHandler *hpal, void *buff,s
     if (ring_buffer_api_is_full(&hpal->tx_queue))
         return PAL_RC_QUEUE_FULL;
 
-    memcpy(hpal->add_to_tx_msg, &size, sizeof(size_t));
-    memcpy(hpal->add_to_tx_msg + sizeof(size_t), buff, size);
+    memcpy(hpal->add_to_tx_msg, &size, sizeof(uint32_t));
+    memcpy(hpal->add_to_tx_msg + sizeof(uint32_t), buff, size);
     RingBufferReturnCode res = ring_buffer_api_push_back(&hpal->tx_queue, hpal->add_to_tx_msg);
     if (res == RING_BUFFER_FULL) {
         return PAL_RC_QUEUE_FULL;
